+++ continuousintegration/orchestration/create_instance.sh
@@ -147,13 +147,13 @@ fi
 poddir=$(mktemp -d)
 
 # Get Catalogue Schema version
-MAJOR=$(grep CTA_CATALOGUE_SCHEMA_VERSION_MAJOR ../../cmake/CTAVersions.cmake | sed 's/[^0-9]*//g')
-MINOR=$(grep CTA_CATALOGUE_SCHEMA_VERSION_MINOR ../../cmake/CTAVersions.cmake | sed 's/[^0-9]*//g')
+MAJOR=$(grep CTA_CATALOGUE_SCHEMA_VERSION_MAJOR ../../catalogue/cta-catalogue-schema/CTACatalogueSchemaVersion.cmake | sed 's/[^0-9]*//g')
+MINOR=$(grep CTA_CATALOGUE_SCHEMA_VERSION_MINOR ../../catalogue/cta-catalogue-schema/CTACatalogueSchemaVersion.cmake | sed 's/[^0-9]*//g')
 SCHEMA_VERSION="$MAJOR.$MINOR"
 
 # It sets as schema version the previous to the current one to create a database with that schema version
 if [ "$updatedatabasetest" == "1" ] ; then
-    MIGRATION_FILE=$(find ../../catalogue/ -name "*to${SCHEMA_VERSION}.sql")
+    MIGRATION_FILE=$(find ../../catalogue/cta-catalogue-schema -name "*to${SCHEMA_VERSION}.sql")
     PREVIOUS_SCHEMA_VERSION=$(echo $MIGRATION_FILE | grep -o -E '[0-9]+\.[0-9]' | head -1)
     NEW_SCHEMA_VERSION=$SCHEMA_VERSION
     SCHEMA_VERSION=$PREVIOUS_SCHEMA_VERSION